Add commit status for E2E Test workflow
Now that the E2E Test workflow is a workflow_run target workflow, it
is disconnected from the source PR in that it does not appear in the
PR's checks list.

We update the workflow to add its own commit status to the PR via the
commit SHA. The context parameter to the create commit status API
connect the two APIs calls so that the later call updates the status
of the initial call.

We also include a target url to the E2E Test action run for the PR.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>